codecatalyst: fix e2e test throttling
Problem:

Our tests are being throttled during the E2E CI

Solution:

Reduce the amount of API requests we make when waiting
for a dev env to delete. Before this commit we checked
if a dev env was deleted every second, but it typically
takes >1 minute to delete.

So this commit increases the amount of time till the next
check from 1 second to 20 seconds.

Signed-off-by: nkomonen <nkomonen@amazon.com>