CodeWhisperer: Fix multiple issues around reporting CodePercentage
1. Add the logic to compute generatedLines count, this will enable
reporting the correct count of lines from CodeWhisperer. Making it
a local computation.
2. Remove the 20 chars check in countTotalTokens(). This will enable
the tracker timer to start whenever there's a change.
3. Remove an additional addTotalToken() call when accepting
CodeWhisperer suggestions, this avoids double counting the total tokens
when accepting CW suggetions.
4. Fix an issue where the referenceCount for each trigger is not reset
after assigned. Removed the global and make it a local computation.
5. Ignore deletion events when computing total tokens, since
it's non trivial to track the deletion of CW suggestions. This implictly
makes the metric to track the code percentage of CW suggestion among
all the code user generates.