auth: change Add Connection page Identity Center submit behavior
Problem:

When there is an error under the submit button of the Identity Center
form for the Add Connection page, the 'sign in' button looks like it
is stuck because we expect the user to update the Identity Center form
(eg: change the start url) and then allow them to click 'sign in'.

The issue is in certain cases where they have an error but want to retry
(eg: cancelled during the sing in process) they are unable to do this.

Solution:

If an error is detected under the 'sign in' button, which is caused
by a previous failed attempt, we will still allow the user to click
'sign in' again.

Signed-off-by: nkomonen <nkomonen@amazon.com>