webview: show error message when webviews throw
Problem:

Before this commit, when webview backend/server code threw
the error would only be logged. So during interactions in
the UI if the user did something and the backend threw
it would not be obvious that something failed.

Solution:

Now, for any error that is thrown we will show an
error message to the user and link to where it failed in the
logs.

Additional:

In some cases, such as show.ts#ssoSetup() of this commit, we
make an effort to handle the exception gracefully, but if it an
unexpected error we will then decide to show the error message

Signed-off-by: nkomonen <nkomonen@amazon.com>