fix: stale log file text when reading
Problem:

We have functionality that shows an error message
to a user and has a button that opens the files and
highlights to the specific error log.

The issue is that on subsequent errors the error
will fail to be highlighted even though we can
see the error in the editor.

Solution:

It looks to be that vscode.TextDocument.getText()
caches the text and only updates once something triggers it.
So when we programatically request the text, it will not match
what we see in the editor.

To fix this, we use the 'revert' command which seems to cause the getText()
to return the latest text content.

Signed-off-by: nkomonen <nkomonen@amazon.com>