fix: show user message on unhandled exception #3868
Problem:
When webview backend/server code throws an unhandled
exception, it is only logged. The UI doesn't expose the message
or any indication about next steps.

Solution:
When a webview throws an unhandled exception, show an error message
with a "View Logs" button.

Problem:
The "View Logs" feature can get confused on subsequent errors,
the error will not be highlighted even though it's in the logs.
Apparently vscode.TextDocument.getText() caches the text and
only updates once something triggers it. So when we programatically
request the text, we sometimes get stale results.

Solution:
Use the vscode "revert" command to force getText() to read from disk.

Signed-off-by: nkomonen <nkomonen@amazon.com>