bug(auth): client registration expiry not checked when creating token
Problem:

In the scenario a client registration was expired and
SsoAccessTokenProvider.createToken() was called, the
expired client registration would be used to create
the new token.

This scenario was not expected to occurr since it was
assumed that getToken() would be called before createToken()
and getToken() would do the work of invalidating the
expired client registration so that createToken() would
not need to check if the client registration is expired.

Solution:

Since we cannot guarantee createToken() is always run
before getToken(), in createToken() we will verify
that the client registration is not expired, and if it is
we will create a new registration, then continue with the same
process as before.

Signed-off-by: nkomonen <nkomonen@amazon.com>