feat(codewhisperer): add direct connect command
Problem:
There is a need to reduce confusion/friction for enterprises who use
IAM Identity Center for CodeWhisperer capability.

At the moment, a customer would need to know that they would need to
follow the "Professional Tier" login flow to use Identity Center.
This may not be obvious for all customers.

Additionally, a customer might already have an active connection to
IAM Identity Center but would need to reauthenticate to update their
scope to include CodeWhisperer.

Focusing on easing this pain for enterprise customers, including Amazon,
there is a benefit to have the ability to directly establish this
connection without forcing enterprise customers to navigate through
the options everytime.

Solution:
Add a new command to directly start the process to connect to
Identity Center, either with predetermined values passed as an
argument or prompt the user to enter the values.