feat(settings): check that user settings.json is valid
Problem:
vscode notifies the user if settings.json is complete nonsense, but is
silent if there are only "recoverable" JSON syntax errors.
So _reading_ settings will appear to work, but _writing_ settings will
throw an exception.

Solution:
On startup, check that we can actually _write_ settings, and show an
error message if not.