telemetry: add additional 'Add Connection' page entrypoints
Problem:

The 'Add Connection' page can be opened from different
places. When we open it from a certain place we indicate
from where. The problem was that certain entrypoints
were missed, resulting in an 'unknown' state.

Solution:

Add the missing entrypoints for the status bar
and explorer when there are no connections at all.

These will be the `statusBar` and `explorer` "source"
value in the metric.

Signed-off-by: nkomonen <nkomonen@amazon.com>