feat(codewhisperer): select customization by prefix
Problem:
The `aws.codeWhisperer.connect` command was introduced to allow enterprises
to establish a default customization for onboarding developers after
establishing an IdC connection.

However, it only selected with a given ARN. Given the current behavior
for adding updated customization for CodeWhisperer, an enterprise would
need to keep the ARN up-to-date as well when calling this command.

Solution:
Provide the ability to define the customization to select with a
prefix for the customization name. This way the logic to select a
customization as updated customizations are introduced can be simplified
on the administrator side by maintaining a prefix.
e.g. MyModel-v1, MyModel-v2, etc.