codewhisperer: refactor create node functions
Problem:
We want to create a quickpick menu for CW with the
same items as the CW tree in Developer Tools.

There is a lot of duplication and it will be easier
to work with if all the implenmentations are located
in a single place.

Solution:

This commit refactors the code to use function overloading
to allow the caller to decide which type of object they
want back (quickpick item vs. tree node).

The next commit will create the implementation for
a quickpick item.

Signed-off-by: nkomonen <nkomonen@amazon.com>