telemetry: include "source" for certain CW commands
Problem:

The same functionality (eg: Pause Suggestions) for CW
can be executed in different places (eg: Tree Node vs
Quick Pick). Our telemetry does not have a way to
indicate this so that we know what triggered the
actual functionality.

Solution:

Include an optional 'source' attribute in the
'vscode_executeCommand' metric so that we can
get this information.

Signed-off-by: nkomonen <nkomonen@amazon.com>