telemetry: separate execute command events with a different 'source'
Problem:

We have a Commands class that allows us to register a
command with vscode, that when called will execute the
callback but more importantly in this PR it emits
a telemetry event.

When a Command is executed in succession within a short
timeframe, the telemetry events emitted are throttled and
a debounce counter is instead incremented. Once the throttle
period is over then a single metric is emitted with that
debounce counter indicating how many times it was called.

The problem is that we throttle command events when emitted,
but if it has different 'source' when executed they all get grouped in to
a single debounce counter. If I execute the same command with
source "a", and another time with source "b" I do not want them
to be merged in to a single debounce counter since the
final metric will only have the 'source' as "a".

Solution:

When a Command is registered or declared, if a 'compositeKey'
is used then it will allow us to differentiate metrics with the
same id by using the specific args to distinguish them.

Additionally these specific args will be used in the fields for
the metric 'vscode_executeCommand'

Signed-off-by: nkomonen <nkomonen@amazon.com>