telemetry: separate execute command events with a different 'source'
Problem:

We have a Commands class that allows us to register a
command with vscode, that when called will execute the
callback but more importantly in this PR it emits
a telemetry event.

When a Command is executed in succession within a short
timeframe, the telemetry events emitted are throttled and
a debounce counter is instead incremented. Once the throttle
period is over then a single metric is emitted with that
debounce counter indicating how many times it was called.

The problem is that we throttle command events when emitted,
but if it has different 'source' when executed they all get grouped in to
a single debounce counter. If I execute the same command with
source "a", and another time with source "b" I do not want them
to be merged in to a single debounce counter since the
final metric will only have the 'source' as "a".

Solution:

If a BaseCommandSource instance is provided in the args
of an RegisteredCommand.execute() call, we will detect
that and treat it as a separate event from the same command
but with a different BaseCommandSource value.

This will result in separate telemetry events each with
their own unique "source" value. It will allow us to reuse
a Command for different entrypoints and still be able to
distinguish it in our telemetry.

Signed-off-by: nkomonen <nkomonen@amazon.com>