refactor: use mapping for auth_addedConnections call
When emitting our telemetry we give it a key which
consists of values separate from AuthFormId. AuthFormId
is good enough and it is better to use that throughout our
code than change it up for a different use case.

This commit uses AuthFormId for the telemetry method
calls, right before telemetry we use a map to convert
the AuthFormId in to the required fields for telemetry.

Signed-off-by: nkomonen <nkomonen@amazon.com>