ci: only run "lint" in "insiders" job
Problem:
The release pipeline is blocked on the lint task because of an
environment issue (there is not actually a git secret in the source
code, this test already passed CI).

  1) git-secrets
       sanity check it finds secrets:
      AssertionError [ERR_ASSERTION]: Expected values to be strictly equal:
      at Context.<anonymous> (src/testLint/gitSecrets.test.ts:125:16)

Solution:
Only run lint in "insiders" job. That allows the release pipeline to
proceed.

Followup to: 6057ef46b7de