telemetry: commands can define how they are deduplicated #3994
Problem:
Our Commands class wraps vscode commands defined in the Toolkit
codebase. The wrapper emits telemetry for each command, but also
_deduplicates_ such telemetry by the command name. For some commands, we
want to consider more than just the command name when deciding if an
invocation is a "duplicate". For example, the `source` that called
a command also should be considered.

Solution:
When a Command is registered or declared, if a `compositeKey` is set,
the Commands deduplicator logic will also use the invocation _arguments_
that map to the `compositeKey` definition, which deciding if the
invocation is unique. These args will also provide values for
corresponding metadata fields on the metric 'vscode_executeCommand'