Q splash page pops on first valid load, move to SVGs (#1247)
Amazon Q Welcome panel now shows up on the first time a valid CWSPR connection is made, and not after that (unless a user reenables the prompt). This will remain accessible through a Command Palette command, although we may move it to an ellipsis command in the Chat panel instead. Note that this now replaces the existing CodeWhisperer Learn panel (which is still accessible through the tree, and is also accessible through the welcome page)

This also moves all iconography to SVGs from PNGs. Since we're using these not as VS Code icons (and instead, as direct images), we will hold off for now on adding them to the icon contribution.