codewhisperer: status bar quickpick when clicked (#3958)
* codewhisperer: refactor create node functions

Problem:
We want to create a quickpick menu for CW with the
same items as the CW tree in Developer Tools.

There is a lot of duplication and it will be easier
to work with if all the implenmentations are located
in a single place.

Solution:

This commit refactors the code to use function overloading
to allow the caller to decide which type of object they
want back (quickpick item vs. tree node).

The next commit will create the implementation for
a quickpick item.

Signed-off-by: nkomonen <nkomonen@amazon.com>

* refactor: remove unused code

Signed-off-by: nkomonen <nkomonen@amazon.com>

* codewhisperer: CW node can give quick pick items as children

Signed-off-by: nkomonen <nkomonen@amazon.com>

* codewhisperer: create quick pick items implementation

Create the implementation for creating a quickpick
item when the user calls the function with the
'type' set to "item".

Signed-off-by: nkomonen <nkomonen@amazon.com>

* codewhisperer: clicking statusbar icon opens quickpick

Signed-off-by: nkomonen <nkomonen@amazon.com>

* codewhisperer: add "sign out" option to quick pick

Signed-off-by: nkomonen <nkomonen@amazon.com>

* codewhisperer: always show status bar icon

Before this commit, we would only show the status bar
icon when the user was logged in.

Now with this commit, we will always show the
status bar icon and when click + user not logged in
they will see the 'Start' option.

Signed-off-by: nkomonen <nkomonen@amazon.com>

* codewhisperer: show reconnect quick pick when expired

Before this commit, when the CW connection was expired
if they clicked the status bar icon they would be immediately
taken through.

Now with this commit, when the status bar icon is clicked
the quickpick will show with the options to re-connect
or sign out.

Signed-off-by: nkomonen <nkomonen@amazon.com>

* fix: circular dependency

- the refreshStatusBar in basicCommands.ts command imported inlineCompletionService.ts
which imported basicCommands.ts

- also the show quick pick command had circular dependency issues since it
  uses the codewhispererNode object. To fix this we moved the command to its
  own module

Signed-off-by: nkomonen <nkomonen@amazon.com>

* test: provided cw quick pick items

Tests the happy path for the expected quickpick
items being provided when CW is connected.

Since we reuse the same node/item generating logic
for the tree we do not need to test all edge cases.

Signed-off-by: nkomonen <nkomonen@amazon.com>

* refactor: improve status bar w/ auto-suggestion icons

Changes:
- Functionality remains the same, except the status bar
  code was moved in to its own class. This is being done to simplify testing

- A "play" and "pause" icon are now shown on the status
  bar to indicate the state of if auto-suggestions are
  enabled.

- Created CodeSuggestionState class to abstract the
  state of if code suggestions are enabled.
  The initial issue was that we didn't have a single
  place to get the latest state and instead needed to
  check the global state. This new class manages that
  completely and is used wherever this state information
  is needed.

Signed-off-by: nkomonen <nkomonen@amazon.com>

* telemetry: include "source" for certain CW commands

Problem:

The same functionality (eg: Pause Suggestions) for CW
can be executed in different places (eg: Tree Node vs
Quick Pick). Our telemetry does not have a way to
indicate this so that we know what triggered the
actual functionality.

Solution:

Include an optional 'source' attribute in the
'vscode_executeCommand' metric so that we can
get this information.

Signed-off-by: nkomonen <nkomonen@amazon.com>

* refactor: fix Add Connection command + telemetry

Signed-off-by: nkomonen <nkomonen@amazon.com>

* fix: getting started page unhandled arg

Since vscode commands can be given incorrect args
when executed, this commit handles that case for
this specific command.

Signed-off-by: nkomonen <nkomonen@amazon.com>

* fix: delay codewhisperer node resolve in initial load

Problem:

A CodeWhispererNode is resolved immediately on module load
and this causes issues since its dependencies may have not
been loaded yet during the extension activation process.

Solution:

Use a getter instead so that we only create the CodeWhispererNode
instance when we need it, and all its dependencies should have
been loaded by then.

Signed-off-by: nkomonen <nkomonen@amazon.com>

* rename: cw quickpick -> cw list commands

Renaming to make naming more generic

Signed-off-by: nkomonen <nkomonen@amazon.com>

* feat: include current model name beside quickpick option

Signed-off-by: nkomonen <nkomonen@amazon.com>

---------

Signed-off-by: nkomonen <nkomonen@amazon.com>