feat: function to get info about all Amazon Q auth state
Problem:

In the Chat UI we need to know about the states of
different auths so that we can show them the appropriate
UI related to the connected state of their auth.

For example, weaverbird only works if IdC is used and
not Builder ID.

Solution:

This commit creates a function getChatAuthState() which when
called will return an object with a status with information
regarding each feature under the CodeWhisperer auth and what
the state of its auth looks like.

For example, if we are using Builder ID, the Amazon Q
entry will have an "unsupported" value since it does not
work with Builder ID.

Additional:

- Existing users of CW will not have the new chat scopes in their
  existing auth connections on we do a full release.
  This commit will allow their existing connection to
  work with those scopes, but if they try and use Chat they
  will need to re-auth.

  In this commit, AuthUtil.reauthenticate additionally adds
  the missing scopes when called. So in states where scopes
  are missing `getChatAuthState()` will indicate those features
  are 'expired' and calling `reauthenticate` will fix those.

Signed-off-by: nkomonen <nkomonen@amazon.com>