fix: vscode command unexpected arg
Problem:

If a command is execute by a VS Code UI component like
an ellipsis or tree node, it is possible for the args
of that component to be incorrect since vscode will not
use the proper arguments.

Solution:

Add a new type called VsCodeCommandArgument that is a
placeholder for that unexpected argument. When the command
is used within the code we should set the value to `undefined`
and when it is executed by a VS Code UI component it will be
defined which indicates the other args will not exist.

Signed-off-by: nkomonen <nkomonen@amazon.com>