CodeWhisperer: fix various VSC bugs
1. Fixed an issue where clientCompoentLatency telemetry event
would be sent every time when navigating the suggestions, where
it should only be sent once when a first suggestion has been shown.
2. Fixed an issue where the requestIdList is not reset, causing a later
check to not pass and stop sending the userTriggerDecisions.
3. Fixed an issue where spamming trigger will cause the
invocation start time to reset to current time for each trigger action.
This will cause the e2e latency to be recorded unexpectedly.