fix: use 99.x as alpha/beta version
Problem:
package.ts generates "alpha" builds using version 1.9999. This was
driven by the notion that we might want all "alpha" testers to be
auto-updated if we ever bump the prod major version. But in practice
there is no clear demarcation where or when that happens. And it means
we need to bump the "alpha" version whenever we bump the prod major
version.

Solution:
Use 99.x for "alpha" builds.

ref 991bfcb794c8699a9ffccdc26b29006a20febe28