Move each Developer Tool to its own view (#1302)
Developer tools show up in their own views. This was done by removing the concept of the `RootNode` (most of which wasn't used) and instead shipping the same node as a tree node, more or less to be used as a container, and emitting the children directly into the tree (the provider just uses the node's children). This will let us use the root node in other places if it ever becomes necessary, and makes it a bit easier to ship views with multiple "root" nodes, in case we have shared functionality we want between them.

<details><summary>Reconnect</summary>

![Screenshot 2023-11-16 082943](https://github.com/aws/aws-toolkit-vscode-staging/assets/29374703/ed4724af-3e59-48bd-913b-301b65fbc0fb)

</details>

<details><summary>Unauthed</summary>

![Screenshot 2023-11-16 083142](https://github.com/aws/aws-toolkit-vscode-staging/assets/29374703/cf25e0fd-f4cf-44f8-9956-dbd66ecf7177)

</details>

<details><summary>Authed</summary>

![Screenshot 2023-11-16 083755](https://github.com/aws/aws-toolkit-vscode-staging/assets/29374703/ad79c521-e84b-470f-9346-75883e7b8649)

</details>