fix: racecondition in cw auth change (#4062)
Problem:

In the Add Connection page we listen to auth
events in the backend. When once of these events
is fired the Add Connection page eventually checks
if isEnterpriseSsoInUse. The issue is that this
uses a variable that is separately updated by
the event onDidChangeActiveConnection.

There is a race condition for when onDidChangeActiveConnection
updates the variable used in isEnterpriseSsoInUse() versus
when isEnterpriseSsoInUse() is actually called.

Also when SecondaryAuth.deleteConnection() is called the
new state is not updated when the function is complete since
we are relying on an event listener to do the final updating.

Solution:

Get rid of the variable and whenever isEnterpriseSsoInUse()
is called it will pull the latest state. Now we don't have
to manage it.

Also in deleteConnection() do a final clearing of the saved
connection state so that upon completion the user will be able
to get the latest state before all the event emitters fire.

Signed-off-by: nkomonen <nkomonen@amazon.com>