Auth: Initial add connection page redesign (#4042)
* authWebview: isConnectionExists method

Adds in a new method for each auth form (eg builder id for CodeWhisperer)
which simply returns a boolean of if a connection for that
form exists (is known by toolkit).

This is being done since we will eventually want to know
for each auth form if a connection exists, since checking
if an auth form connection is actively connected does not
tell us if the connection exists. It may exist but not be
active.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* refactor: auth webview to use existing auths at start

Previously, we only looked at actively connected auths
when taking the initial snapshot.

Now, we will instead look at all existing auths that
the extension is aware of. This will handle edge cases
where a connection exists, but is not connected and that
does not get considered as an existing connection.

Additionally, renames some variables + minor refactor

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* fix: disabling credentials submit button appropriately

There was some old logic the enabled/disabled the submit
button. This commit removes it.

Additionally, there was a slight refactor to appropriately
use the submit button disable logic. What it now does is
disables the submit button while the authentication of the
credentials is happening. Once a result is returned, regardless
of success, we enable the button again.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* refactor: Add error messages to state instead of frontend

In the credentials auth form we manage error messages in
the frontend code.

This moves the error message management in to the state class
since this is a better design.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* remove unused code

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* refactor: credentials in to state

This refactors the credentials logic all in to the
state class so that we can eventually extract the state
class in to the backend code and properly test it.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* refactor: builder id use state

Improves how we manage state for Builder ID

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* refactor: identity center use state more

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* fix: credentials content panel not loading

This was not loading due to a race condition which
is described in a comment in the code of this commit.

- Additionally, slightly refactor updateIsAllAuthsLoaded() to
  improve readability.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* test/refactor: auth webview code + tests for emitWebviewClosed()

Signed-off-by: nkomonen <nkomonen@amazon.com>

* refactor: use mapping for auth_addedConnections call

When emitting our telemetry we give it a key which
consists of values separate from AuthFormId. AuthFormId
is good enough and it is better to use that throughout our
code than change it up for a different use case.

This commit uses AuthFormId for the telemetry method
calls, right before telemetry we use a map to convert
the AuthFormId in to the required fields for telemetry.

Signed-off-by: nkomonen <nkomonen@amazon.com>

* tests: add + fix tests for auth_addedConnections

Signed-off-by: nkomonen <nkomonen@amazon.com>

* refactor: remove unnecessary code

Signed-off-by: nkomonen <nkomonen@amazon.com>

* refactor: initial auth count is collapsed in telemetry

now when we show the count of the initial
auths it will be collapsed. eg: not every
idc and credential are counted. Instead they
are collapsed in to their own count.

Signed-off-by: nkomonen <nkomonen@amazon.com>

* refactor: notification messages

Better architect the notification messages.

- Move the definitions of them in to their own components
- Create a Model+Controller class to manage them and consolidate
  all the notification related code in to a single class.

TODO: Look to dynamically create notifications so we can simply add
      a new notification implementation and can call the controller
      to create it.

      The difficult part is getting type completion for the different
      types of notifications since their Props can vary and we'll want
      a nice way to provide an ID then get the proper type suggestions
      for the correct args.

Signed-off-by: nkomonen <nkomonen@amazon.com>

* remove stuff

Signed-off-by: nkomonen <nkomonen@amazon.com>

* refactor: webview scaling is reactive to width

Signed-off-by: nkomonen <nkomonen@amazon.com>

* working: before screwing with scaling

Signed-off-by: nkomonen <nkomonen@amazon.com>

* refactor: sizing/scaling is handled

Signed-off-by: nkomonen <nkomonen@amazon.com>

* refactor: All feature panels are aligned now

Signed-off-by: nkomonen <nkomonen@amazon.com>

* refactor: auth changes in backend update frontend

This needed to be fixed due to the overall frontend
refactor

Signed-off-by: nkomonen <nkomonen@amazon.com>

* refactor: various fixes

Signed-off-by: nkomonen <nkomonen@amazon.com>

* feat: open feedback form button in Add Connection page

Signed-off-by: nkomonen <nkomonen@amazon.com>

* refactor: improve UI refreshing when auth changes

- Now refresh the individual feature instead of all
  of them since we have more granularity in to what
  changed

Signed-off-by: nkomonen <nkomonen@amazon.com>

* fix: first time user error on auto open add connection page

Since the Auth module was loaded before telemetry, when auth
ran it emitted a telemetry event before telemetry was set up.

This has telemetry setup first, then auth.

Signed-off-by: nkomonen <nkomonen@amazon.com>

* fix: circular dependency issues

Moves code around to resolve circular dependency issues

Signed-off-by: nkomonen <nkomonen@amazon.com>

* webview: improve refreshing/loading of auth forms

Less stuttery and more isolated

Signed-off-by: nkomonen <nkomonen@amazon.com>

* changelog item

Signed-off-by: nkomonen <nkomonen@amazon.com>

* fix unit tests

Signed-off-by: nkomonen <nkomonen@amazon.com>

* codecatalyst: minor UI fix in add connection page

Signed-off-by: nkomonen <nkomonen@amazon.com>

* refactor: improve reloading/rendering

Will keep the Identity Center forms exposed
if they are connected.

Signed-off-by: nkomonen <nkomonen@amazon.com>

* refactor: add in separate CW manage connections command

As to not disturb the downstream private repo I am
adding this back in for now

Signed-off-by: nkomonen <nkomonen@amazon.com>

---------

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>
Signed-off-by: nkomonen <nkomonen@amazon.com>