fix: CW/CC auth not properly saved
Problem:

In `SecondaryAuth.useNewConnection()` we use this method when there is a
new connection that we want to associate with the specific feature (CC/CW).

The issue is that the current implementation of `useNewConnection()` will not
"save" that connection for the specific feature, and when the IDE is reloaded
the feature will not have a pointer to the previously configured connection
that it can automatically connect to.

Solution:

Change how `useNewConnection()` works so that we will always save the
connection to the specific feature when it is added, but also allow
users of the primary "Auth" class to use it if no available connection
exists.

Signed-off-by: nkomonen <nkomonen@amazon.com>