fix: Refresh connection state before checking chat state
Problem:

When we call `getChatAuthState()` we are relying on the
main Auth to have properly update the state of the Auth
so that we can get an accurate result.

The problem is that there is no guarantee that the connection
state up to date when `getChatAuthState()` is called.

Solution:

In `getChatAuthState()` we call `refreshConnectionState()` to
update our state store with the latest connection state. It is
essentially checking if a token can be retrieved and updating
the store that indicates if it can.

As a side-effect, all usages of `getChatAuthStates()` must be
updated to work `async` since `getChatAuthState()` had to be
changed to `async`

Signed-off-by: nkomonen <nkomonen@amazon.com>