fix: connect to existing builder id if it has necessary scopes
Problem:

If we try to connect to a CC builder id but it already
exists locally we fail.

This scenario happenend in the Add Connection page when
we were connected to Builder ID, switched to IdC, then
attempted to connect back to Builder ID. The Builder ID
connection existed but would not be used and caused a
failure.

Solution:

If we are trying to connect to Builder ID and it exists
with the correct scopes already, then return that.

Signed-off-by: nkomonen <nkomonen@amazon.com>