Fix(Starling) - Reject responses from CWChat with CodeReferences
Problem:
AppComposer does not have a workflow to detect when and where a user accepts a suggestion. This is required to build up the log entry that goes into the CodeWhisperer Reference Log.

Solution:
Starling currently only supports single CFN resource prompts. After a large analysis, we have found that the CW Chat model does not generate responses that contain code references to these style of prompts.

Therefore, for now, we are opting to reject any responses that may have a code reference until we have the capability needed to fully track a users acceptance to specific lines in their template and correctly update the reference log.