amazonq: show badge on Q icon for existing CodeWhisperer users
Problem:
When Q is launched the icon may not be noticed.

Solution:
If the user has previously used Codewhisperer, show a "1" badge on the
Q icon. The intention is for a user to click the view, get told they
need to reauth into Codewhisperer to get new tokens, and propel them
into the flow.

This does NOT show for users that have not previously tried
Codewhisperer.