fix: "Command aws.amazonq.refresh not found"
Problem:
CodeWhisperer can be enabled in environments where AmazonQ is not. But
'aws.amazonq.refresh' command is called in many codewhisperer "refresh"
routines, and if AmazonQ was not enabled this causes an error:

    aws.auth.manageConnections: Error: Command 'aws.amazonq.refresh' not found.

Solution:
Introduce `tryExecute()`. This should probably be a common practice
since there are many places where we can't assume that a command exists.

Alternatively we could "forward declare" all commands, which may exist
only as stubs if their implementation is not activated.