fix(awsq): add "Enable Amazon Q" button for connections lacking q chat scopes (#1529)
Problem: "Switch to Amazon Q Chat" will display if the user has a valid existing CW connection but
it does not have Q scopes (e.g. upgrading to an Amazon Q toolkit verison). This implies that Q
chat is ready but in will in fact fail without the scopes.

Solution: Check for proper scopes and if not, re-authenticate to get correct scopes. Display a
more informative message.