fix(auth): preserve old scopes if re-auth fails. show more helpful error message
Problem: If a connection fails due to not having valid scopes, the scopes are not reset. This
soft-locks the user and new connections aren't possible until it is deleted in the quick pick.

Solution: Restore scopes if that is the failure reason. Also, add a more helpful UI message to
indicate why the sign on failed.