refactor: do validation checks AFTER user picks project #4093
Problem:
We run a shell command (`javap`) to determine the Java version of each of the
projects open in the user's workspace. Running this command for each open
project is totally unnecessary.

Solution:
Show options to transform any of the projects open in the workspace, and then
once the user has selected a project, the validation checks are run (`javap`
etc.) to check that it is a valid project. The solution is mostly a refactoring
-- the logic to do the validation checks was pre-existing, it is simply moved to
happen after the user selects a project.

Co-authored-by: David Hasani <davhasan@amazon.com>