Fix(AppComposer): Enable sync's lambda validation check to be skipped
Problem:
The Sam sync command enforces* that a Serverless Function exist in the template else it will throw a hard error. This requirement seems unnecessary as the SAM CLI has no such requirement.

Solution:
Expose a new, optional, argument to the sam sync command (`validate`) - by default, if not defined, it will maintain the current behavior. If the `validate` argument is defined, it will be passed through to `CloudFormation.load`, which already supported this functionality.