refactor: remove yaml dependency, use js-yaml instead.
Problem: 'yaml' is outdated and requires human intervention to migrate to V2. It is only used in a few places in the codebase, whereas
everywhere else, a similar yaml parser js-yaml is used.

Solution: remove usage of 'yaml' and only use 'js-yaml', as it has some features that we need (ability to write out to a string and
skip invalid JSON entries, which 'yaml' does not seem to offer). We do not require all the features of 'yaml' at this time. Also,
remove "errors" field from the state machines parser -> webview message. It is not used by anything and was only accessible with
'yaml'.

In one case of 'yaml', we used the 'maxAliasCount' option to prevent users from submitting state machines with aliases in them.
Aliases do not prevent publishing normally to step functions if used normally; I believe it is to prevent the user from breaking the
toolkit by exponentially copying anchors/aliases in the yaml. This does not affect publishing (it will fail as it is supposed to if
the defintion is too large), and slightly slows down toolkit usage while the parse is happening. This validation is not necessary because this sort of
behavior would be very specific and intentional from the user, and in my testing still causes slow downs and errors due to the logic
centered around how the validation option was used.