codecatalyst: Bad auth state when adding connection + skip onboarding
Problem:

When a user connects to CC but skips the prompt to onboard a space, things appear to fail to the user even though things are fine. They will see a cancelled error under the Builder ID button and a "Sign in" node on the CodeCatalyst tree view.

Solution:

We ignore the error that gets raised if they cancel and successfully add the connection.

Also if they require onboarding we will show them the "You must onboard" node instead of the incorrect "You must sign in" node that we are currently showing. When they click this node it will open the CodeCatalyst page in the browser that guides them to onboard.

Signed-off-by: nkomonen <nkomonen@amazon.com>