amazonQ: only add missing scopes intentionally
Problem:

Before we would always add missing scopes whenever an
existing CW user was missing the new scopes for chat.
This would be done during the reauthentication process
automatically.

The issue is that we did not want this and only wanted
to add missing Amazon Q/chat scopes when the user intentionally
wanted them to be added.

Solution:
- Don't always add missing scopes during reauthenticate(). Instead
  use an arg to determine if we should
- We will add missing scopes in the Chat "reauthenticate" prompt as well
  as the Amazon Q node.
- Everywhere else will simply reauthenticate without adding new scopes

Signed-off-by: nkomonen <nkomonen@amazon.com>